Fixed im_to_rgb

+++ datasets/clusters_to_instances.py
@@ -6,6 +6,8 @@ Created on Sat Sep 14 20:53:38 2019
 @author: sumche
 """
 import numpy as np
+import matplotlib.pyplot as plt
+
 
 class Point():
     def __init__(self, x_dist, y_dist, row, col):
@@ -155,12 +157,14 @@ def get_color(num):
 def to_rgb(bw_im):
     instances = np.unique(bw_im)
     instances = instances[instances != 0]
-    rgb_im = [np.zeros(bw_im.shape), np.zeros(bw_im.shape), np.zeros(bw_im.shape)]
+    rgb_im = [np.zeros(bw_im.shape, dtype=int), 
+              np.zeros(bw_im.shape, dtype=int), 
+              np.zeros(bw_im.shape, dtype=int)]
     for instance in instances:
         color = get_color(instance)
-        rgb_im[0][instance == bw_im] = int(color[0])
-        rgb_im[1][instance == bw_im] = int(color[1])
-        rgb_im[2][instance == bw_im] = int(color[2])
+        rgb_im[0][instance == bw_im] = color[0]
+        rgb_im[1][instance == bw_im] = color[1]
+        rgb_im[2][instance == bw_im] = color[2]
     return np.stack([rgb_im[0],rgb_im[1],rgb_im[2]],axis=-1)
     #return np.concatenate([np.concatenate([np.expand_dims(rgb_im[0], -1), np.expand_dims(rgb_im[1], -1)], 2), np.expand_dims(rgb_im[2], -1)], 2)
